+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- fix(sumologicexporter): treat resource attributes as fields for otlp #536
+
 [Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.47.0-sumo-0...main
+[#536]: https://github.com/SumoLogic/sumologic-otel-collector/pull/536
 
 ## [v0.47.0-sumo-0]